pyproject.toml
[tool.poetry]
name = "volttron"
version = "10.1.0"
description = "VOLTTRON™ is an open source platform for distributed sensing and control. The platform provides services for collecting and storing data from buildings and devices and provides an environment for developing applications which interact with that data."
authors = ["volttron <volttron@pnnl.gov>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://volttron.org"
repository = "https://github.com/eclipse-volttron/volttron-core"
documentation = "https://eclipse-volttron.readthedocs.io"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: Implementation :: CPython"
]

packages = [
    { include = "volttron", from = "src" },
    { include = "corefixtures", from = "tests" }
]

[tool.poetry.dependencies]
poetry = "^1.2.2"
python = "^3.10"
gevent = "^23.9.1"
PyYAML = "^6.0.1"
toml = "^0.10.2"
dateutils = "^0.6.12"
tzlocal = "^5.2"
psutil = "^5.9.8"
cryptography = "^41.0.7"
watchdog-gevent = "^0.1.1"
deprecated = "^1.2.14"
dataclass-wizard = "^0.22.2"
pip = "^24.0"
returns = "^0.22.0"
coloredlogs = "^15.0.1"
blinker = "^1.7.0"

[tool.poetry.group.dev.dependencies]
volttron-test-utils = { path="../volttron-test-utils", develop = true}

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
volttron = 'volttron.server.__main__:main'
vcfg = 'volttron.client.commands.config:main'
volttron-cfg = 'volttron.client.commands.config:main'
vctl = 'volttron.client.commands.control:main'
volttron-ctl = 'volttron.client.commands.control:main'

[tool.yapfignore]
ignore_patterns = [
    ".venv/**",
    ".pytest_cache/**",
    "dist/**",
    "docs/**"
]

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
column_limit = 120
split_before_logical_operator = true

[tool.mypy]
plugins = ["returns.contrib.mypy.returns_plugin"]
show_error_context = true
pretty = true
show_column_numbers = true
warn_return_any = true
warn_unused_configs = true
allow_redefinition = false
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
local_partial_types = true
no_implicit_optional = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true
